[depthMap] CUDA: Move host/common files
