+++ src/aliceVision/depthMap/ROI.hpp
