+++ src/aliceVision/depthMap/cuda/host/DeviceCamera.cpp
