+++ src/aliceVision/depthMap/cuda/host/DeviceStreamManager.cpp
