+++ src/aliceVision/depthMap/cuda/host/DeviceStreamManager.hpp
