+++ src/aliceVision/depthMap/cuda/imageProcessing/deviceGaussianFilter.hpp
@@ -6,8 +6,8 @@
 
 #pragma once
 
-#include <aliceVision/depthMap/cuda/memory.hpp>
-#include <aliceVision/depthMap/cuda/device/device_utils.h>
+#include <aliceVision/depthMap/BufPtr.hpp>
+#include <aliceVision/depthMap/cuda/host/memory.hpp>
 
 #include <set>
