+++ src/aliceVision/depthMap/cuda/normalMapping/DeviceNormalMapper.cpp
@@ -6,7 +6,7 @@
 
 #include "DeviceNormalMapper.hpp"
 
-#include <aliceVision/depthMap/cuda/memory.hpp>
+#include <aliceVision/depthMap/cuda/host/memory.hpp>
 
 namespace aliceVision {
 namespace depthMap {
