+++ src/aliceVision/depthMap/cuda/normalMapping/deviceNormalMap.cu
@@ -6,10 +6,10 @@
 
 #include "deviceNormalMap.hpp"
 
-#include <aliceVision/depthMap/cuda/memory.hpp>
-#include <aliceVision/depthMap/cuda/hostUtils.hpp>
+#include <aliceVision/depthMap/cuda/host/memory.hpp>
+#include <aliceVision/depthMap/cuda/host/hostUtils.hpp>
 #include <aliceVision/depthMap/cuda/device/matrix.cuh>
-#include <aliceVision/depthMap/cuda/device/utils.cuh>
+#include <aliceVision/depthMap/cuda/device/buffer.cuh>
 #include <aliceVision/depthMap/cuda/normalMapping/eig33.cuh>
 
 #include <math_constants.h>
