+++ src/aliceVision/depthMap/cuda/planeSweeping/deviceDepthSimilarityMap.cu
@@ -7,7 +7,7 @@
 #include "deviceDepthSimilarityMap.hpp"
 #include "deviceDepthSimilarityMapKernels.cuh"
 
-#include <aliceVision/depthMap/cuda/hostUtils.hpp>
+#include <aliceVision/depthMap/cuda/host/hostUtils.hpp>
 
 #include <utility>
