+++ src/aliceVision/depthMap/cuda/planeSweeping/deviceDepthSimilarityMapKernels.cuh
@@ -6,8 +6,8 @@
 
 #pragma once
 
-#include <aliceVision/depthMap/cuda/ROI.hpp>
-#include <aliceVision/depthMap/cuda/device/utils.cuh>
+#include <aliceVision/depthMap/ROI.hpp>
+#include <aliceVision/depthMap/cuda/device/buffer.cuh>
 #include <aliceVision/depthMap/cuda/device/matrix.cuh>
 #include <aliceVision/depthMap/cuda/device/Patch.cuh>
