+++ src/aliceVision/depthMap/cuda/planeSweeping/deviceSimilarityVolume.hpp
@@ -8,10 +8,10 @@
 
 #include <aliceVision/depthMap/SgmParams.hpp>
 #include <aliceVision/depthMap/RefineParams.hpp>
+#include <aliceVision/depthMap/ROI.hpp>
 
-#include <aliceVision/depthMap/cuda/memory.hpp>
-#include <aliceVision/depthMap/cuda/DeviceCamera.hpp>
-#include <aliceVision/depthMap/cuda/ROI.hpp>
+#include <aliceVision/depthMap/cuda/host/memory.hpp>
+#include <aliceVision/depthMap/cuda/host/DeviceCamera.hpp>
 #include <aliceVision/depthMap/cuda/planeSweeping/similarity.hpp>
 
 namespace aliceVision {
