+++ src/aliceVision/depthMap/cuda/planeSweeping/deviceSimilarityVolumeKernels.cuh
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <aliceVision/depthMap/cuda/ROI.hpp>
+#include <aliceVision/depthMap/ROI.hpp>
 #include <aliceVision/depthMap/cuda/device/matrix.cuh>
 #include <aliceVision/depthMap/cuda/device/Patch.cuh>
 #include <aliceVision/depthMap/cuda/planeSweeping/similarity.hpp>
