+++ src/aliceVision/depthMap/volumeIO.hpp
@@ -11,7 +11,7 @@
 #include <aliceVision/depthMap/SgmParams.hpp>
 #include <aliceVision/depthMap/RefineParams.hpp>
 #include <aliceVision/depthMap/DepthSimMap.hpp>
-#include <aliceVision/depthMap/cuda/memory.hpp>
+#include <aliceVision/depthMap/cuda/host/memory.hpp>
 #include <aliceVision/depthMap/cuda/planeSweeping/similarity.hpp>
 
 #include <string>
